Initial test setup
Setup the home page tests, url, view and templates.  Commented out the Flask template tags in the original main.html and renamed it to base.html. All of the Flask template tags need to be converted to Django.

Added a if block to the root urls to handle django debug toolbar urls when running in development.